Move touch from core.util to core.path
Move its tests, too.
